Add hyperlink injection grammar.
This is aggressively injected into the text, string, and comment scopes. We may want to restrict the text scope so that we don’t markup hyperlinks in text formatting languages where links require special markup, so as to not confuse the user (about whether his links are auto-detected during conversion).