Add ‘issue’ as a new scheme
I am not aware of any precedent for this, but I wish to make it easy to reference issues in the code base without giving a full URL to the issue tracker. With project specific variables we can place the URL (prefix/template) in a project variable and magically have all the issue:// links (for that project) work.