Limit time/size of download to prevent long waits
Limit the download time to four seconds: This functions to prevent waiting for a large file to download, it also prevents waiting for a slow server or a broken link. This won't actually prevent a large download from being started but will prevent it from having much of an impact.

Also updates the script to ruby 2. Updating the script and adding the timeout necessitated changing the `TM_LINK_FORMAT` that use the title value. This is an unfortunate change as users may have created their own versions but in the worst case an empty title value is created. (The title variable will now always exist but with a nil value when there is no title found.)